データベースのマイグレーションを整理し、不要なテーブルを削除。ユーザー管理のための新しいテーブルを追加し、ユーザー名を使用するように変更。
