+++ app/api/login/route.ts
@@ -1,22 +1,22 @@
 import { NextResponse } from "next/server";
 import { db } from "../../../drizzle/db";
-import { user_settings } from "../../../drizzle/schema";
+import { user } from "../../../drizzle/schema";
 import { eq } from "drizzle-orm";
 
 export async function POST(req: Request) {
-  const { userId } = await req.json();
-  if (!userId) {
+  const { username } = await req.json();
+  if (!username) {
     return NextResponse.json(
-      { ok: false, error: "userId必須" },
+      { ok: false, error: "username必須" },
       { status: 400 }
     );
   }
   const exists = await db
     .select()
-    .from(user_settings)
-    .where(eq(user_settings.user_id, userId));
+    .from(user)
+    .where(eq(user.username, username));
   if (exists.length > 0) {
-    return NextResponse.json({ ok: true });
+    return NextResponse.json({ ok: true, userId: exists[0].id });
   } else {
     return NextResponse.json(
       { ok: false, error: "ユーザーが存在しません" },
