+++ drizzle/schema.ts
@@ -1,4 +1,11 @@
-import { pgSchema, serial, text, varchar, jsonb } from "drizzle-orm/pg-core";
+import {
+  pgSchema,
+  serial,
+  text,
+  varchar,
+  jsonb,
+  integer,
+} from "drizzle-orm/pg-core";
 
 // スキーマを定義
 export const schema = pgSchema("notify_app");
@@ -6,7 +13,7 @@ export const schema = pgSchema("notify_app");
 // スキーマ内にテーブルを定義
 export const records = schema.table("records", {
   id: serial("id").primaryKey(),
-  user_id: varchar("user_id", { length: 64 }).notNull(),
+  user_id: integer("user_id").notNull(),
   date: varchar("date", { length: 32 }).notNull(), // ISO8601文字列で保存
   mood: jsonb("mood").notNull(), // string[]
   diary: text("diary").notNull(),
@@ -15,15 +22,21 @@ export const records = schema.table("records", {
 // 目標・習慣テーブル
 export const goals = schema.table("goals", {
   id: serial("id").primaryKey(),
-  user_id: varchar("user_id", { length: 64 }).notNull(),
+  user_id: integer("user_id").notNull(),
   habit: text("habit").notNull(),
   goal: text("goal").notNull(),
 });
 
-// ユーザー設定テーブル
-export const user_settings = schema.table("user_settings", {
+// ユーザーテーブル
+export const user = schema.table("user", {
   id: serial("id").primaryKey(),
-  user_id: varchar("user_id", { length: 64 }).notNull(),
+  username: varchar("username", { length: 64 }).notNull().unique(),
+});
+
+// 通知設定テーブル
+export const notify_settings = schema.table("notify_settings", {
+  id: serial("id").primaryKey(),
+  user_id: integer("user_id").notNull(),
   type: varchar("type", { length: 32 }).notNull(),
   custom_message: text("custom_message").notNull(),
 });
@@ -32,14 +45,14 @@ export const user_settings = schema.table("user_settings", {
 export const subscriptions = schema.table("subscriptions", {
   id: serial("id").primaryKey(),
   endpoint: text("endpoint").notNull().unique(),
-  user_id: varchar("user_id", { length: 64 }),
+  user_id: integer("user_id"),
   keys: jsonb("keys").notNull(), // {p256dh, auth}
 });
 
 // AIログテーブル
 export const ai_logs = schema.table("ai_logs", {
   id: serial("id").primaryKey(),
-  user_id: varchar("user_id", { length: 64 }).notNull(),
+  user_id: integer("user_id").notNull(),
   timestamp: varchar("timestamp", { length: 32 }).notNull(),
   prompt: text("prompt").notNull(),
   response: text("response").notNull(),
